Proyecto2/ApiPython/requirements.txt
Flask==2.3.3
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
mysql-connector-python==8.2.0
redis==5.0.1